Bundle cjs and esm with rollup (#22)
In this diff I added rollup to bundle cjs and esm. With flat bundles we
can easier to control the size user consumes.

size snapshot plugin allows to track that size with every commit.
Moreover it shows treeshakability of the lib for esm bundle.

It can be less critical for user bundle but make sense when the lib is
used by another one with a set of components. So one unused component
depending on this lib will not keep this dependency in user bundle.